Refactored routes and templates to use wtforms

+++ app/routes.py
@@ -1,6 +1,7 @@
-from flask import Blueprint, render_template, redirect, request, flash
+from flask import Blueprint, render_template, redirect, request, flash, url_for
 from .models import User, Post, Tag, PostTag, db
 from sqlalchemy.exc import SQLAlchemyError
+from .forms import NewUserForm, NewPostForm, NewTagForm
 
 main = Blueprint('main', __name__)
 
@@ -28,21 +29,24 @@ def show_user_directory():
 @user.route("/users/new")
 def show_create_user_form():
     """Show form to add a user"""
-    return render_template("users/new.html")
+    form = NewUserForm()
+
+    return render_template("users/new.html", form=form)
 
 
 @user.route("/users/new", methods=["POST"])
 def add_user():
     """Add user to users db with user input values"""
-    first_name = request.form["first_name"]
-    last_name = request.form["last_name"]
-    image_url = request.form["image_url"]
+    form = NewUserForm()
+    first_name = form.first_name.data
+    last_name = form.last_name.data
+    image_url = form.image_url.data
 
     new_user = User(first_name=first_name,
                     last_name=last_name, image_url=image_url)
     db.session.add(new_user)
     db.session.commit()
-    return redirect("/users")
+    return redirect(url_for('user.show_user_directory'))
 
 
 @user.route("/users/<int:user_id>")
@@ -63,14 +67,15 @@ def edit_user_form(user_id):
 @user.route("/users/<int:user_id>/edit", methods=["POST"])
 def edit_user(user_id):
     """Show form to edit a user"""
+    form = NewUserForm()
     user = User.query.filter_by(id=user_id).first()
 
-    user.last_name = request.form["last_name"]
-    user.first_name = request.form["first_name"]
-    user.image_url = request.form["image_url"]
+    user.last_name = form.last_name.data
+    user.first_name = form.first_name.data
+    user.image_url = form.image_url.data
     db.session.commit()
 
-    return redirect("/users")
+    return redirect(url_for('user.show_user_directory'))
 
 
 @user.route("/users/<int:user_id>/delete", methods=["POST"])
@@ -79,7 +84,7 @@ def delete_user(user_id):
     User.query.filter_by(id=user_id).delete()
     db.session.commit()
 
-    return redirect("/users")
+    return redirect(url_for('user.show_user_directory'))
 
 
 # ===================POSTS ROUTES===================
@@ -89,37 +94,44 @@ post = Blueprint('post', __name__)
 @post.route("/users/<int:user_id>/posts/new")
 def show_post_form(user_id):
     """Show a form to add a post for that user"""
+    form = NewPostForm()
+
     user = User.query.get_or_404(user_id)
     tags = Tag.query.all()
-    return render_template("posts/new.html", user=user, tags=tags)
+
+    return render_template("posts/new.html", form=form, user=user, tags=tags)
 
 
 @post.route("/users/<int:user_id>/posts/new", methods=["POST"])
 def add_new_post(user_id):
     """Handle add form, add post and redirect to user's details page"""
+    form = NewPostForm()
     user = User.query.get_or_404(user_id)
-    print(f"THIS SHOULD BE THE USER ID: {user.id}")
-    new_post = Post(title=request.form["post_title"],
-                    content=request.form["post_content"], user_id=user.id)
 
-    try:
-        db.session.add(new_post)
-        db.session.flush()  # make sure new_post has an ID before committing
+    if form.validate_on_submit():
+        new_post = Post(title=form.title.data,
+                        content=form.content.data, user_id=user.id)
 
-        tags = request.form.getlist("tag")
-        print(tags)
+        try:
+            db.session.add(new_post)
+            db.session.flush()
 
-        for tag in tags:
-            tag = Tag.query.filter_by(name=tag).first()
-            new_post_tag = PostTag(post_id=new_post.id, tag_id=int(tag.id))
-            db.session.add(new_post_tag)
+            tags = request.form.getlist("tag")
 
-        db.session.commit()
-    except SQLAlchemyError as e:
-        print(str(e))
-        db.session.rollback()
+            for tag_name in tags:
+                tag = Tag.query.filter_by(name=tag_name).first()
+                if tag:
+                    new_post_tag = PostTag(
+                        post_id=new_post.id, tag_id=int(tag.id))
+                    db.session.add(new_post_tag)
 
-    return redirect(f"/users/{user.id}")
+            db.session.commit()  # Commit changes after all tags are processed
+
+        except SQLAlchemyError as e:
+            print(str(e))
+            db.session.rollback()
+
+    return redirect(url_for('user.show_user', user_id=user_id))
 
 
 @post.route("/posts/<int:post_id>")
@@ -135,35 +147,34 @@ def show_post(post_id):
 @post.route("/posts/<int:post_id>/edit")
 def edit_post_form(post_id):
     """Show form to edit a post"""
+    form = NewPostForm()
     post = Post.query.get_or_404(post_id)
     user = User.query.get_or_404(post.user_id)
     tags = Tag.query.all()
 
-    return render_template("/posts/edit.html", user=user, post=post, tags=tags)
+    return render_template("/posts/edit.html", user=user, post=post, tags=tags, form=form)
 
 
 @post.route("/posts/<int:post_id>/edit", methods=["POST"])
 def edit_post(post_id):
     """Handle editing of a post, redirect back to post detail view."""
+    form = NewPostForm()
     post = Post.query.get_or_404(post_id)
 
-    post.title = request.form["post_title"]
-    post.content = request.form["post_content"]
+    if form.validate_on_submit():
+        post.title = form.title.data
+        post.content = form.content.data
 
-    tag_names = request.form.getlist("tags")
-    print(tag_names)
+        tag_names = request.form.getlist("tags")
+        tags = Tag.query.filter(Tag.name.in_(tag_names)).all()
+        post.tags = tags
 
-    tags = Tag.query.filter(Tag.name.in_(tag_names)).all()
-    post.tags = tags
+        db.session.commit()
 
-    print("++++++++++++++++++++++++++++++++")
-    print(tags)
-    print(post.tags)
-    print("++++++++++++++++++++++++++++++++")
-
-    db.session.commit()
-
-    return redirect(f"/posts/{post_id}")
+        return redirect(url_for('post.show_post', post_id=post_id))
+    else:
+        flash('Form submission failed. Please check your input.', 'error')
+        return redirect(url_for('post.edit_post_form', post_id=post_id))
 
 
 @post.route("/posts/<int:post_id>/delete", methods=["POST"])
@@ -175,7 +186,7 @@ def delete_post(post_id):
     Post.query.filter_by(id=post_id).delete()
     db.session.commit()
 
-    return redirect(f"/users/{user.id}")
+    return redirect(url_for('post.show_post', post_id=post_id))
 
 
 # ==============TAG ROUTES====================
@@ -186,7 +197,6 @@ tag = Blueprint('tag', __name__)
 def show_all_tags():
     """Lists all tags with links to each tag's detail page"""
     tags = Tag.query.all()
-    print(tags)
 
     return render_template("/tags/alltags.html", tags=tags)
 
@@ -196,52 +206,73 @@ def show_tag_details(tag_id):
     """Show tag's detail page"""
     tag = Tag.query.get_or_404(tag_id)
     tagged_posts = tag.posts
-    print(f"Tag: {tag}")
-    print(f"Tagged posts: {tagged_posts}")
+
     return render_template("/tags/details.html", tag=tag, tagged_posts=tagged_posts)
 
 
 @tag.route("/tags/new")
 def show_create_tag_form():
     """Show a form to create a tag"""
-    return render_template("/tags/new.html")
+    form = NewTagForm()
+
+    return render_template("/tags/new.html", form=form)
 
 
 @tag.route("/tags/new", methods=["POST"])
 def create_tag():
     """Create a tag, add to Tag table"""
-    new_tag = Tag(name=request.form["tag_name"])
-    db.session.add(new_tag)
-    db.session.commit()
+    form = NewTagForm()
+    if form.validate_on_submit():
+        new_tag = Tag(name=form.name.data)
 
-    return redirect("/tags")
+        db.session.add(new_tag)
+        db.session.commit()
+
+        return redirect(url_for('tag.show_all_tags'))
+    else:
+        flash('Form submission failed. Please check your input.', 'error')
+        return render_template("/tags/new.html", form=form)
 
 
 @tag.route("/tags/<int:tag_id>/edit")
 def show_edit_tag_form(tag_id):
     """Show a form to edit a tag"""
-
+    form = NewTagForm()
     tag = Tag.query.get_or_404(tag_id)
 
-    return render_template("/tags/edit.html", tag=tag)
+    return render_template("/tags/edit.html", tag=tag, form=form)
 
 
 @tag.route("/tags/<int:tag_id>/edit", methods=["POST"])
 def edit_tag(tag_id):
     """Process edit form, update tag in db, redirect to tag list"""
+    form = NewTagForm()
     tag = Tag.query.get_or_404(tag_id)
-    tag.name = request.form["tag_name"]
 
-    db.session.commit()
-    return redirect("/tags")
+    if form.validate_on_submit():
+        tag.name = form.name.data
+        print(tag.name)
 
-# 7. To-do: POST route /tags/[tag-id]/delete = Delete a tag
+        db.session.commit()
+        return redirect(url_for('tag.show_all_tags'))
+    else:
+        flash('Form submission failed. Please check your input.', 'error')
+        return render_template("/tags/edit.html", form=form, tag=tag)
 
 
 @tag.route("/tags/<int:tag_id>/delete", methods=["POST"])
 def delete_tag(tag_id):
     """Delete tag, update db"""
-    Tag.query.filter_by(id=tag_id).delete()
-
-    db.session.commit()
-    return redirect("/tags")
+    tag = Tag.query.get_or_404(tag_id)
+    print(tag)
+    try:
+        if tag:
+            Tag.query.filter_by(id=tag_id).delete()
+            db.session.commit()
+        else:
+            flash("Tag not found.", "error")
+    except Exception as e:
+        print(e)  # print exception for debugging
+        db.session.rollback()
+        flash('Delete failed.', 'error')
+    return redirect(url_for('tag.show_all_tags'))
